+++ src/routes/notification.js
@@ -5,5 +5,6 @@ const ctrl = require('../controllers/notificationController');
 
 router.get('/', ctrl.list);
 router.put('/:id/read', ctrl.markRead);
+router.put('/read-all', ctrl.markAllRead);
 
 module.exports = router;